% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQmetrics_example_data.r
\docType{data}
\name{evidence}
\alias{evidence}
\title{evidence.txt}
\format{
data.frame
}
\usage{
data("MQmetrics_example_data")
}
\description{
The evidence file combines all the information about the identified peptides
and normally is the only file  required for processing the results.
}
\keyword{data.frame}
