% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentInformation.R
\name{ExperimentInformation}
\alias{ExperimentInformation}
\title{Experiment Duration}
\usage{
ExperimentInformation(running_times, parameters)
}
\arguments{
\item{running_times}{The runningTimes.txt table from  MaxQuant Output.}

\item{parameters}{The parameters.txt table from  MaxQuant Output.}
}
\value{
Returns the time in hours:minutes that lasted the whole Experiment.
}
\description{
Experiment Duration
}
