% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Genearte a report with all functions combined}
\usage{
generateReport(input_dir, dir = getwd())
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Genearte a report with all functions combined
}
