% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MQCombined.R
\name{make_MQCombined}
\alias{make_MQCombined}
\title{Read MaxQuant Tables From Directory}
\usage{
make_MQCombined(MQPathCombined, remove_contaminants = TRUE)
}
\arguments{
\item{MQPathCombined}{The directory to the "combined" folder where the
MaxQuant results are stored.}

\item{remove_contaminants}{Whether or not to remove contaminants,
reverse and identified by one one peptide.}
}
\value{
The files from the MaxQuant with the contaminants and Reverse
hits removed.
}
\description{
Read MaxQuant Tables From Directory
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
MQCombined <- make_MQCombined(MQPathCombined)
}
