% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQmetrics_example_data.r
\docType{data}
\name{modificationSpecificPeptides}
\alias{modificationSpecificPeptides}
\title{modificationSpecificPeptides.txt}
\format{
data.frame
}
\usage{
data("MQmetrics_example_data")
}
\description{
modificationSpecificPeptides.txt
}
\keyword{data.frame}
