% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQPathCombined.r, R/MQmetrics_example_data.r
\docType{data}
\name{MQPathCombined}
\alias{MQPathCombined}
\title{Directory to the MaxQuant Output created automatically named combined.}
\format{
Value.

Value.
}
\usage{
data("MQmetrics_example_data")

data("MQmetrics_example_data")
}
\description{
In this combined foulder should be at least two folders: txt and proc.

In this combined foulder should be at least two folders: txt and proc.
}
\keyword{Value}
