% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQmetrics_example_data.r
\docType{data}
\name{msScans}
\alias{msScans}
\title{msScans.txt}
\format{
data.frame
}
\usage{
data("MQmetrics_example_data")
}
\description{
The msScans table contains information about the full scans, which can be used to verify data quality and
generated useful statistics about the interaction between the samples and LC.
}
\keyword{data.frame}
