% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAcquisitionCycle.R
\name{PlotAcquisitionCycle}
\alias{PlotAcquisitionCycle}
\title{Acquisition Cycle and MS/MS}
\usage{
PlotAcquisitionCycle(MQCombined, palette = "Set2", plots_per_page = 5)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Two plots per sample, one with the cycle tyme vs retention time,
 and MS/MS count vs retention time.
}
\description{
Acquisition Cycle and MS/MS
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
MQCombined <- make_MQCombined(MQPathCombined)
PlotAcquisitionCycle(MQCombined)

}
