% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAllDynamicRange.R
\name{PlotAllDynamicRange}
\alias{PlotAllDynamicRange}
\title{Plots the dynamic range for all samples}
\usage{
PlotAllDynamicRange(proteinGroups, show_shade = TRUE, percent_proteins = 0.9)
}
\arguments{
\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}

\item{show_shade}{Creates a shade showing where the \code{percent_proteins} are.
Default is TRUE.}

\item{percent_proteins}{Determines the percentage for the show_shade parameter.
Default is 0.90 (90\% of the proteins).}
}
\value{
Returns one plot for each sample, being the dynamic range.
}
\description{
Plots the dynamic range for all samples
}
\examples{
MQPathCombined <- system.file('extdata', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
proteinGroups <- files[["proteinGroups.txt"]]
PlotAllDynamicRange(proteinGroups)

}
