% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAndromedaScore.R
\name{PlotAndromedaScore}
\alias{PlotAndromedaScore}
\title{Andromeda score for the best associated MS/MS spectrum.}
\usage{
PlotAndromedaScore(
  MQCombined,
  show_median = TRUE,
  size_median = 1.5,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{show_median}{If true it will show the median of each group, as a red
dashed line.By default is TRUE.}

\item{size_median}{The width of the median line in the plots.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Plots the MaxQuant Andromeda Score.
}
\description{
Andromeda score for the best associated MS/MS spectrum.
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotAndromedaScore(MQCombined)
}
