% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCharge.R
\name{PlotCharge}
\alias{PlotCharge}
\title{The charge-state of the precursor ion.}
\usage{
PlotCharge(
  MQCombined,
  palette = "Set2",
  plots_per_page = 5,
  tabular_output = FALSE
)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Plots the charge-state of the precursor ion.
}
\description{
The charge-state of the precursor ion.
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotCharge(MQCombined)
}
