% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCombinedDynamicRange.R
\name{PlotCombinedDynamicRange}
\alias{PlotCombinedDynamicRange}
\title{Dynamic range of all the samples combined}
\usage{
PlotCombinedDynamicRange(MQCombined, show_shade = TRUE, percent_proteins = 0.9)
}
\arguments{
\item{show_shade}{Creates a shade showing where the \code{percent_proteins}
are. Default is TRUE.}

\item{percent_proteins}{Determines the percentage for the show_shade
parameter. Default is 0.90 (90\% of the proteins).}

\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}
}
\value{
Returns the dynamic range for all samples combined.
}
\description{
Dynamic range of all the samples combined
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
proteinGroups <- files[['proteinGroups.txt']]
PlotCombinedDynamicRange(proteinGroups)
}
