% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIdentifactionRatio.R
\name{PlotIdentificationRatio}
\alias{PlotIdentificationRatio}
\title{Identification Ratio Between Peptides and Proteins}
\usage{
PlotIdentificationRatio(
  MQCombined,
  intensity_type = "Intensity",
  long_names = FALSE,
  sep_names = NULL
)
}
\arguments{
\item{intensity_type}{}
}
\value{

}
\description{
Identification Ratio Between Peptides and Proteins
}
