% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIdentificationType.R
\name{PlotIdentificationType}
\alias{PlotIdentificationType}
\title{Protein and Peptide Identification type.}
\usage{
PlotIdentificationType(
  MQCombined,
  long_names = FALSE,
  sep_names = NULL,
  palette = "Set2"
)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{long_names}{If TRUE, samples having long names will be considered, and
the name will be split by sep_names. By default = FALSE.}

\item{sep_names}{If long_names is TRUE, sep_names has to be selected. Samples
names will be split. By default is NULL.}

\item{palette}{The palette from the Package RColorBrewer. By default
is 'Set2'.}
}
\value{
Plots the compares of the protein and peptide  identification
of each sample. It will only work if in MaxQuant the Match
Between Run was selected.
}
\description{
Protein and Peptide Identification type.
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotIdentificationType(MQCombined)
}
