% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIntensity.R
\name{PlotIntensity}
\alias{PlotIntensity}
\title{Intensity / LFQ intensity per sample.}
\usage{
PlotIntensity(
  proteinGroups,
  split_violin_intensity = TRUE,
  intensity_type = "Intensity",
  log_base = 2,
  long_names = FALSE,
  sep_names = NULL,
  palette = "Set2"
)
}
\arguments{
\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}

\item{split_violin_intensity}{If TRUE, both the LFQ and the Intensity will be
shown in the same plot. If FALSE, it can be specified in the intensity_type
which intensity to visualize.}

\item{intensity_type}{The type of intensity. Values: 'Intensity' or 'LFQ'.
Only useful if split_violin_intensity = FALSE.  Default is Intensity.}

\item{log_base}{The logarithmic scale for the intensity. Default is 2.}

\item{long_names}{If TRUE, samples having long names will be considered, and
the name will be split by sep_names. By default = FALSE.}

\item{sep_names}{If long_names is TRUE, sep_names has to be selected. Samples
names will be split. By default is NULL.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
A violin plot and boxplot of the intensities in each sample.
}
\description{
Intensity / LFQ intensity per sample.
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
proteinGroups <- files[['proteinGroups.txt']]
PlotIntensity(proteinGroups)


}
