% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIsotopePattern.R
\name{PlotIsotopePattern}
\alias{PlotIsotopePattern}
\title{Plot Isotope Pattern and Isotope Pattern Sequenced}
\usage{
PlotIsotopePattern(summary, position_dodge_width = 1, font_size = 12)
}
\arguments{
\item{summary}{The summary.txt table from  MaxQuant Output.}

\item{position_dodge_width}{Width overlapping columns.}

\item{font_size}{Size of the font in the labels.}
}
\value{
Returns a plot Isotope Pattern and Isotope Pattern Sequenced.
}
\description{
Plot Isotope Pattern and Isotope Pattern Sequenced
}
