% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMissedCleavages.R
\name{PlotMissedCleavages}
\alias{PlotMissedCleavages}
\title{Missed Cleavages}
\usage{
PlotMissedCleavages(peptides, position_dodge_width = 1)
}
\arguments{
\item{peptides}{peptides.txt table from MaxQuant ouput.}

\item{position_dodge_width}{position of the columns within each others.}
}
\value{
A plot showing the missed cleavages in each column.
}
\description{
Missed Cleavages
}
