% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMsMs.R
\name{PlotMsMs}
\alias{PlotMsMs}
\title{Comparison of the MS/MS submmited and identified in each sample.}
\usage{
PlotMsMs(summary, position_dodge_width = 1, font_size = 12)
}
\arguments{
\item{summary}{The summary.txt table from  MaxQuant Output.}

\item{position_dodge_width}{Width overlapping columns.}

\item{font_size}{Size of the font in the labels.}
}
\value{
Plots the MS/MS submited and Identified in each sample.
}
\description{
Comparison of the MS/MS submmited and identified in each sample.
}
