% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPeaks.R
\name{PlotPeaks}
\alias{PlotPeaks}
\title{Total number of peaks detected and sequenced}
\usage{
PlotPeaks(
  summary,
  long_names = FALSE,
  sep_names = NULL,
  position_dodge_width = 1,
  palette = "Set2"
)
}
\arguments{
\item{summary}{The summary.txt table from  MaxQuant Output.}

\item{long_names}{If TRUE, samples having long names will be considered, and
the name will be split by sep_names. By default = FALSE.}

\item{sep_names}{If long_names is TRUE, sep_names has to be selected. Samples
names will be split. By default is NULL.}

\item{position_dodge_width}{Position of the columns within each others.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Plots the total number of peaks detected in the full scans and the
 total number of peaks sequenced by tandem MS.
}
\description{
Total number of peaks detected and sequenced
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
summary <- files[['summary.txt']]
PlotPeaks(summary)

}
