% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteaseSpecificity.R
\name{PlotProteaseSpecificity}
\alias{PlotProteaseSpecificity}
\title{Protease Specificity}
\usage{
PlotProteaseSpecificity(peptides, position_dodge_width = 0.2, palette = "Set2")
}
\arguments{
\item{peptides}{peptides.txt table from MaxQuant ouput.}

\item{position_dodge_width}{position of the columns within each others.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Two plots per sample: Peptide length distribution and the number of
 missed enzymatic cleavages.
}
\description{
Protease Specificity
}
\examples{
MQPathCombined <- '/home/alvaro/Documents/MaxQuant/example4/'
files <- ReadDataFromDir(MQPathCombined)
peptides <- files[['peptides.txt']]
PlotProteaseSpecificity(peptides)

}
