% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinCoverageAll.R
\name{PlotProteinCoverageAll}
\alias{PlotProteinCoverageAll}
\title{Protein Coverage of all samples.}
\usage{
PlotProteinCoverageAll(proteinGroups)
}
\arguments{
\item{proteinGroups}{proteinGroups.txt table from MaxQuant output.}
}
\value{
A plot showing the protein coverage in all samples.
}
\description{
Protein Coverage of all samples.
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
proteinGroups <- files[['proteinGroups.txt']]
PlotProteinCoverageAll(proteinGroups)

}
