% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinOverlap.R
\name{PlotProteinOverlap}
\alias{PlotProteinOverlap}
\title{Protein Overlap Between Samples}
\usage{
PlotProteinOverlap(MQCombined)
}
\arguments{
\item{proteinGroups}{proteinGroups.txt table from MaxQuant output.}
}
\value{
A plot showing the protein coverage in all samples.
}
\description{
Protein Overlap Between Samples
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
proteinGroups <- files[['proteinGroups.txt']]
PlotProteinOverlap(proteinGroups)

}
