% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinsIdentified.R
\name{PlotProteinsIdentified}
\alias{PlotProteinsIdentified}
\title{Title}
\usage{
PlotProteinsIdentified(
  proteinGroups,
  intensity_type = "Intensity",
  long_names = FALSE,
  sep_names = NULL,
  palette = "Set2",
  font_size = 12
)
}
\arguments{
\item{proteinGroups}{}
}
\value{

}
\description{
Title
}
