% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTM.R
\name{PlotPTM}
\alias{PlotPTM}
\title{Title}
\usage{
PlotPTM(modificationSpecificPeptides, peptides_modified = 5, palette = "Set2")
}
\arguments{
\item{modificationSpecificPeptides}{}
}
\value{

}
\description{
Title
}
