% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTotalIonCurrent.R
\name{PlotTotalIonCurrent}
\alias{PlotTotalIonCurrent}
\title{Total Ion Current}
\usage{
PlotTotalIonCurrent(
  MQCombined,
  show_max_value = TRUE,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{show_max_value}{If TRUE, it will show the max TIC value of each sample.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Returns a plot the Total Ion Current in each sample.
The maximum value is also plotted.
}
\description{
Total Ion Current
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotTotalIonCurrent(MQCombined)
}
