% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQmetrics_example_data.r
\docType{data}
\name{proteinGroups}
\alias{proteinGroups}
\title{proteinGroups.txt}
\format{
data.frame
}
\usage{
data("MQmetrics_example_data")
}
\description{
The Protein Groups table contains information on the identified proteins in the processed raw-files. Each
single row contains the group of proteins that could be reconstructed from a set of peptides.
}
\keyword{data.frame}
