% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataFromDir.R
\name{ReadDataFromDir}
\alias{ReadDataFromDir}
\title{Read MaxQuant Tables From Directory}
\usage{
ReadDataFromDir(x, remove_contaminants = TRUE)
}
\arguments{
\item{x}{The directory containing the MaxQuant Output.}

\item{remove_contaminants}{Whether or not to remove contaminants, reverse and identified by one one peptide.}
}
\value{
The output from \code{ReadDataFromR}
}
\description{
Read MaxQuant Tables From Directory
}
