% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTables.R
\name{ReportTables}
\alias{ReportTables}
\title{Report Tables with summary data}
\usage{
ReportTables(
  MQCombined,
  long_names = FALSE,
  sep_names = NULL,
  log_base = 2,
  intensity_type = "Intensity"
)
}
\arguments{
\item{MQCombined}{The directory to the "combined" folder where the
MaxQuant results are stored.}

\item{long_names}{If TRUE, samples having long names will be considered, and
the name will be split by sep_names. By default = FALSE.}

\item{sep_names}{If long_names is TRUE, sep_names has to be selected. Samples
names will be split. By default is NULL.}

\item{log_base}{The logarithmic scale for the intensity. Default is 2.}

\item{intensity_type}{The type of intensity. Values: 'Intensity' or 'LFQ'.}
}
\value{
A list with four tables are generated:
 - Protein Information
 - Intensity Information
 - Peptide Charge Information
 - Peptide hydrophobicity Information
}
\description{
Report Tables with summary data
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
MQCombined <- make_MQCombined(MQPathCombined)
ReportTables(MQCombined)
}
