% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQmetrics_example_data.r
\docType{data}
\name{runningTimes}
\alias{runningTimes}
\title{#runningTimes.txt}
\format{
data.frame
}
\usage{
data("MQmetrics_example_data")
}
\description{
#runningTimes.txt
}
\keyword{data.frame}
