% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQmetrics_example_data.r
\docType{data}
\name{summary}
\alias{summary}
\title{summary.txt}
\format{
data.frame
}
\usage{
data("MQmetrics_example_data")
}
\description{
The summary file contains summary information for all the raw files processed with a single MaxQuant run.
The summary information consists of some MaxQuant parameters, information of the raw file contents, and
statistics on the peak detection. Based on this file a quick overview can be gathered on the quality of the data
in the raw file.
The last row in this file contains the summary information for each column on each of the processed files.
}
\keyword{data.frame}
