% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentInformation.R
\name{ExperimentInformation}
\alias{ExperimentInformation}
\title{Experiment Information}
\usage{
ExperimentInformation(runningTimes, parameters)
}
\arguments{
\item{runningTimes}{The runningTimes.txt table from  MaxQuant Output.}

\item{parameters}{The parameters.txt table from  MaxQuant Output.}
}
\value{
Returns the time in hours:minutes that lasted the whole Experiment.
}
\description{
Experiment Information
}
\examples{
MQPathCombined <- '/home/alvaro/Documents/MaxQuant/example4/'
files <- ReadDataFromDir(MQPathCombined)
runningTimes <-  files[["#runningTimes.txt"]]
parameters <- files[["parameters.txt"]]
ExperimentInformation(runningTimes, parameters)

}
