% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Genearte a report with all functions combined}
\usage{
generateReport(
  input_dir,
  output_dir = getwd(),
  log_base = 2,
  UniprotID = NULL,
  long_names = FALSE,
  sep_names = NULL,
  intensity_type = "Intensity",
  palette = "Set2"
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Genearte a report with all functions combined
}
