% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Genearte a report with all functions combined}
\usage{
generateReport(
  MQPathCombined,
  output_dir = getwd(),
  report_tables = FALSE,
  log_base = 2,
  long_names = FALSE,
  sep_names = NULL,
  intensity_type = "Intensity",
  palette = "Set2",
  UniprotID = NULL,
  segment_width = 1,
  show_shade = TRUE,
  percent_proteins = 0.9,
  show_calibrated_rt = FALSE,
  show_max_value = TRUE,
  peptides_modified = 3
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Genearte a report with all functions combined
}
