% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQuantAnalysisInfo.R
\name{MaxQuantAnalysisInfo}
\alias{MaxQuantAnalysisInfo}
\title{Experiment Information}
\usage{
MaxQuantAnalysisInfo(MQCombined)
}
\arguments{
\item{MQPathCombined}{The directory to the "combined" folder where the
MaxQuant results are stored.}

\item{runningTimes}{The runningTimes.txt table from  MaxQuant Output.}

\item{parameters}{The parameters.txt table from  MaxQuant Output.}
}
\value{
Returns the time in hours:minutes that lasted the whole Experiment.
}
\description{
Experiment Information
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
parameters <- files[["parameters.txt"]]
runningTimes <-  files[["#runningTimes.txt"]]
MaxQuantAnalysisInfo(MQPathCombined, runningTimes, parameters)

}
