% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAcquisitionCycle.R
\name{PlotAcquisitionCycle}
\alias{PlotAcquisitionCycle}
\title{Acquisition Cycle and MS/MS}
\usage{
PlotAcquisitionCycle(msScans, palette = "Set2", plots_per_page = 5)
}
\arguments{
\item{msScans}{The msScans.txt file from the MaxQuant ouptut.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Two plots per sample, one with the cycle tyme vs retention time,
 and MS/MS count vs retention time.
}
\description{
Acquisition Cycle and MS/MS
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
msScans <- files[['msScans.txt']]
PlotAcquisitionCycle(msScans)

}
