% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAllDynamicRange.R
\name{PLotAllDynamicRange}
\alias{PLotAllDynamicRange}
\title{Plots the dynamic range for all samples}
\usage{
PLotAllDynamicRange(proteinGroups, columns = 1, rows = 1)
}
\arguments{
\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}

\item{columns}{The number of column to show the output.}

\item{rows}{The number of rows to show the output.}
}
\value{
Returns one plot for each sample, being the dynamic range. Besides, it can shows where 90\% of the proteins are and their orders of abundance.
}
\description{
Plots the dynamic range for all samples
}
