% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAndromedaScore.R
\name{PlotAndromedaScore}
\alias{PlotAndromedaScore}
\title{Andromeda score for the best associated MS/MS spectrum.}
\usage{
PlotAndromedaScore(peptides, palette = "Set2", plots_per_page = 5)
}
\arguments{
\item{peptides}{peptides.txt table from MaxQuant ouput.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Plots the MaxQuant Andromeda Score.
}
\description{
Andromeda score for the best associated MS/MS spectrum.
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
peptides <- files[['peptides.txt']]
PlotAndromedaScore(peptides)
}
