% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAndromedaScore.R
\name{PlotAndromedaScore}
\alias{PlotAndromedaScore}
\title{Andromeda score for the best associated MS/MS spectrum.}
\usage{
PlotAndromedaScore(peptides, palette = "Set2")
}
\arguments{
\item{peptides}{}
}
\value{

}
\description{
Andromeda score for the best associated MS/MS spectrum.
}
