% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAndromedaScore.R
\name{PlotAndromedaScore}
\alias{PlotAndromedaScore}
\title{Andromeda score for the best associated MS/MS spectrum.}
\usage{
PlotAndromedaScore(peptides, palette = "Set2")
}
\arguments{
\item{peptides}{peptides.txt table from MaxQuant ouput.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Plots the MaxQuant Andromeda Score.
}
\description{
Andromeda score for the best associated MS/MS spectrum.
}
\examples{
data("MQmetrics_example_data")
PlotAndromedaScore(peptides)
}
