% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCharge.R
\name{PlotCharge}
\alias{PlotCharge}
\title{The charge-state of the precursor ion.}
\usage{
PlotCharge(evidence, palette = "Set2")
}
\arguments{
\item{evidence}{The evidence.txt file from the MaxQuant ouptut.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Plots the charge-state of the precursor ion.
}
\description{
The charge-state of the precursor ion.
}
\examples{
files <- ReadDataFromDir(MQPathCombined)
evidence <- files[['evidence.txt']]
PlotCharge(evidence)


}
