% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCharge.R
\name{PlotCharge}
\alias{PlotCharge}
\title{The charge-state of the precursor ion.}
\usage{
PlotCharge(evidence)
}
\arguments{
\item{evidence}{The evidence.txt file from the MaxQuant ouptut.}
}
\value{
Plots the charge-state of the precursor ion.
}
\description{
The charge-state of the precursor ion.
}
