% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCharge.R
\name{PlotCharge}
\alias{PlotCharge}
\title{The charge-state of the precursor ion.}
\usage{
PlotCharge(MQCombined, palette = "Set2", plots_per_page = 5)
}
\arguments{
\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}

\item{evidence}{The evidence.txt file from the MaxQuant ouptut.}
}
\value{
Plots the charge-state of the precursor ion.
}
\description{
The charge-state of the precursor ion.
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
evidence <- files[['evidence.txt']]
PlotCharge(evidence)
}
