% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCombinedDynamicRange.R
\name{PlotCombinedDynamicRange}
\alias{PlotCombinedDynamicRange}
\title{Plots the dynamic range of all the samples combined}
\usage{
PlotCombinedDynamicRange(proteinGroups)
}
\arguments{
\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}
}
\value{
Returns the dynamic range for all samples combined. Besides, it can shows where 90\% of the proteins are and their orders of abundance.
}
\description{
Plots the dynamic range of all the samples combined
}
