% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHydrophobicity.R
\name{PlotHydrophobicity}
\alias{PlotHydrophobicity}
\title{Peptide hydrophobicity by GRAVY score}
\usage{
PlotHydrophobicity(
  peptides,
  show_median = TRUE,
  binwidth = 0.2,
  palette = "Set2"
)
}
\arguments{
\item{peptides}{The peptides file from the MaxQuant ouptut.}

\item{show_median}{If true it will show the median of each group, as a red
dashed line.By default is TRUE.}

\item{binwidth}{Selects the binwidth of the histogram. By default = 0.2}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Returns a histogram per sample, showing the frequency of the peptide's
 hydrophobicity GRAVY value.
}
\description{
Peptide hydrophobicity by GRAVY score
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
peptides <- files[['peptides.txt']]
PlotHydrophobicity(peptides)

}
