% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIdentificationType.R
\name{PlotIdentificationType}
\alias{PlotIdentificationType}
\title{Compares the type of identification of each sample}
\usage{
PlotIdentificationType(
  peptides,
  proteinGroups,
  font_size = 12,
  long_names = FALSE,
  sep_names = "-",
  palette = "Set2"
)
}
\arguments{
\item{peptides}{The peptides.txt table from  MaxQuant Output.}
}
\value{
Plots the compares of the type of identification of each sample. It will not work if in MaxQuant the Match Between Run was not selected.
}
\description{
Compares the type of identification of each sample
}
