% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIntensity.R
\name{PlotIntensity}
\alias{PlotIntensity}
\title{Intensity / LFQ intensity per sample.}
\usage{
PlotIntensity(proteinGroups, intensity_type = "Intensity", log_base = 10)
}
\arguments{
\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}

\item{intensity_type}{The type of intensity. Values: 'Intensity', 'LFQ'.}

\item{log_base}{The logarithmic scale for the intensity. Values: '10', '2', 'none'.}
}
\value{
A violin plot of the intensities in each sample.
}
\description{
Intensity / LFQ intensity per sample.
}
