% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotiRT.R
\name{PlotiRT}
\alias{PlotiRT}
\title{Max intensities of the iRT peptides in each sample.}
\usage{
PlotiRT(MQCombined, show_calibrated_rt = FALSE, tolerance = 0.001)
}
\arguments{
\item{show_calibrated_rt}{If TRUE, it will also show the calibrated retention
time of each iRT peptide. By default = FALSE.}

\item{tolerance}{Error maximum to find the iRT peptides by m/z value.
by default is 0.001.}

\item{evidence}{evidence.txt table from the MaxQuant ouptut.}
}
\value{
A plot showing the iRT peptide in each sample vs the Retention time.
}
\description{
Max intensities of the iRT peptides in each sample.
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
evidence <- files[['evidence.txt']]
PlotiRT(evidence)

}
