% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotiRT.R
\name{PlotiRT}
\alias{PlotiRT}
\title{Max intensities of the iRT peptides in each sample.}
\usage{
PlotiRT(evidence, tolerance = 0.001)
}
\arguments{
\item{evidence}{evidence.txt table from the MaxQuant ouptut.}
}
\value{
A plot showing the iRT peptide in each sample vs the Retention time.
}
\description{
Max intensities of the iRT peptides in each sample.
}
