% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotiRTScore.R
\name{PlotiRTScore}
\alias{PlotiRTScore}
\title{Score vs retention time of the iRT peptides}
\usage{
PlotiRTScore(MQCombined, tolerance = 0.001, plots_per_page = 5)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{tolerance}{Error maximum to find the iRT peptides by m/z value.
by default is 0.001.}

\item{plots_per_page}{Establish the maximum number of plots per page.#'}
}
\value{
A plot for each sample showing a linear regression of the
iRT peptides'
 retention time vs the score.
}
\description{
Score vs retention time of the iRT peptides
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotiRT(MQCombined)
}
