% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotiRTScore.R
\name{PlotiRTScore}
\alias{PlotiRTScore}
\title{Score vs retention time of the iRT peptides}
\usage{
PlotiRTScore(evidence, tolerance = 0.001)
}
\arguments{
\item{evidence}{evidence.txt file from the MaxQuant output.}

\item{tolerance}{Error maximum to find the iRT peptides by m/z value.
by default is 0.001.}
}
\value{
A plot for each sample showing a linear regression of the iRT peptides'
 retention time vs the score.
}
\description{
Score vs retention time of the iRT peptides
}
\examples{
data("MQmetrics_example_data")
PlotiRT(evidence)

}
