% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCA.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Principal Component Analysis of the Intensity values.}
\usage{
PlotPCA(proteinGroups, intensity_type = "Intensity", palette = "Set2")
}
\arguments{
\item{proteinGroups}{The proteinGroups.txt table from  MaxQuant Output.}

\item{intensity_type}{The type of intensity. Values: 'Intensity' or 'LFQ'.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.
Default is Intensity.}
}
\value{
A PCA plot of the Intesities of all the samples.
}
\description{
Principal Component Analysis of the Intensity values.
}
\examples{
MQPathCombined <- '/home/alvaro/Documents/MaxQuant/example4/'
files <- ReadDataFromDir(MQPathCombined)
proteinGroups <- files[['proteinGroups.txt']]
PlotPCA(proteinGroups)

}
