% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCA.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Principal Component Analysis of the Intensity values.}
\usage{
PlotPCA(MQCombined, intensity_type = "Intensity", palette = "Set2")
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{intensity_type}{The type of intensity. Values: 'Intensity' or 'LFQ'.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.
 Default is Intensity.}
}
\value{
A PCA plot of the Intesities of all the samples.
}
\description{
Principal Component Analysis of the Intensity values.
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotPCA(MQCombined)
}
