% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPeaks.R
\name{PlotPeaks}
\alias{PlotPeaks}
\title{Total number of peaks detected and sequenced}
\usage{
PlotPeaks(summary, position_dodge_width = 1, font_size = 12)
}
\arguments{
\item{summary}{The summary.txt table from  MaxQuant Output.}

\item{position_dodge_width}{Width overlapping columns.}

\item{font_size}{Size of the font in the labels.}
}
\value{
Plots the total number of peaks detected in the full scans and the total number of peaks sequenced by tandem MS.
}
\description{
Total number of peaks detected and sequenced
}
