% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPeptidesIdentified.R
\name{PLotPeptidesIdentified}
\alias{PLotPeptidesIdentified}
\title{Total number of peaks detected and sequenced}
\usage{
PLotPeptidesIdentified(peptides, long_names = FALSE, sep_names = "-")
}
\arguments{
\item{peptides}{The peptides.txt table from  MaxQuant Output.}
}
\value{
Plots the total number of unique peptide amino acid sequences identified from the recorded tandem mass spectra.
}
\description{
Total number of peaks detected and sequenced
}
