% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPeptidesIdentified.R
\name{PlotPeptidesIdentified}
\alias{PlotPeptidesIdentified}
\title{Total number of peaks detected and sequenced}
\usage{
PlotPeptidesIdentified(
  MQCombined,
  long_names = FALSE,
  sep_names = NULL,
  palette = "Set2"
)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{long_names}{If TRUE, samples having long names will be considered, and
the name will be split by sep_names. By default = FALSE.}

\item{sep_names}{If long_names is TRUE, sep_names has to be selected. Samples
names will be split. By default is NULL.}

\item{palette}{The palette from the Package RColorBrewer. By default
is 'Set2'.}
}
\value{
Plots the total number of unique peptide amino acid sequences
identified from the recorded tandem mass spectra.
}
\description{
Total number of peaks detected and sequenced
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
MQCombined <- make_MQCombined(MQPathCombined)
PlotPeptidesIdentified(MQCombined)

}
