% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteaseSpecificity.R
\name{PlotProteaseSpecificity}
\alias{PlotProteaseSpecificity}
\title{Protease Specificity}
\usage{
PlotProteaseSpecificity(MQCombined, palette = "Set2", plots_per_page = 5)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Two plots per sample: Peptide length distribution and the number of
 missed enzymatic cleavages.
}
\description{
Protease Specificity
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
MQCombined <- make_MQCombined(MQPathCombined)
PlotProteaseSpecificity(MQCombined)

}
