% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteaseSpecificity.R
\name{PlotProteaseSpecificity}
\alias{PlotProteaseSpecificity}
\title{Protease Specificity}
\usage{
PlotProteaseSpecificity(peptides, palette = "Set2", plots_per_page = 5)
}
\arguments{
\item{peptides}{peptides.txt table from MaxQuant ouput.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}

\item{position_dodge_width}{position of the columns within each others.}
}
\value{
Two plots per sample: Peptide length distribution and the number of
 missed enzymatic cleavages.
}
\description{
Protease Specificity
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
peptides <- files[['peptides.txt']]
PlotProteaseSpecificity(peptides)

}
