% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinCoverage.R
\name{PlotProteinCoverage}
\alias{PlotProteinCoverage}
\title{Protein coverage and degradation.}
\usage{
PlotProteinCoverage(
  peptides,
  proteinGroups,
  UniprotID = NULL,
  log_base = 2,
  segment_width = 1,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{peptides}{peptides.txt table from MaxQuant ouput.}

\item{proteinGroups}{proteinGroups.txt table from MaxQuant output.}

\item{UniprotID}{Uniprot ID of the protein of interest.}

\item{log_base}{The logarithmic scale for the intensity. Default is 2.}

\item{segment_width}{Width of the segments to improve visualization.
Default is 1.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Two plots for each sample, the end position vs the start position of
 each peptide of the given protein found. And the Intensity of a given
 peptide and its length.
}
\description{
Protein coverage and degradation.
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
peptides <- files[['peptides.txt']]
proteinGroups <- files[['proteinGroups.txt']]
PlotProteinCoverage(peptides, proteinGroups, UniprotID = 'Q8R0Y6')

}
