% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinCoverage.R
\name{PlotProteinCoverage}
\alias{PlotProteinCoverage}
\title{Protein coverage and degradation}
\usage{
PlotProteinCoverage(
  peptides,
  proteinGroups,
  UniprotID = NULL,
  log_base = 2,
  segment_width = 1
)
}
\arguments{
\item{peptides}{}
}
\value{

}
\description{
Protein coverage and degradation
}
