% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinOverlap.R
\name{PlotProteinOverlap}
\alias{PlotProteinOverlap}
\title{Protein Overlap Between Samples}
\usage{
PlotProteinOverlap(MQCombined)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}
}
\value{
A plot showing the protein coverage in all samples.
}
\description{
Protein Overlap Between Samples
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotProteinOverlap(MQCombined)
}
