% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTM.R
\name{PlotPTM}
\alias{PlotPTM}
\title{Post Translational Modifications}
\usage{
PlotPTM(
  modificationSpecificPeptides,
  peptides_modified = 1,
  plot_unmodified_peptides = FALSE,
  log_base = 2,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{modificationSpecificPeptides}{modificationSpecificPeptides table from MaxQuant ouput.}

\item{peptides_modified}{Minimum number of peptides modified. Default  is 5.}

\item{plot_unmodified_peptides}{If TRUE, it will show the Unmodified peptides.}

\item{log_base}{The logarithmic scale for the intensity. Default is 2.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Two plots per sample
}
\description{
Post Translational Modifications
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
modificationSpecificPeptides <- files[['modificationSpecificPeptides.txt']]
PlotPTM(modificationSpecificPeptides)

}
