% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTM.R
\name{PlotPTM}
\alias{PlotPTM}
\title{Post Translational Modifications}
\usage{
PlotPTM(
  MQCombined,
  peptides_modified = 1,
  plot_unmodified_peptides = FALSE,
  log_base = 2,
  aggregate_PTMs = TRUE,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{MQCombined}{Object list containing all the files from the MaxQuant
output. It is the result from using \code{make_MQCombined}.}

\item{peptides_modified}{Minimum number of peptides modified. Default  is 5.}

\item{plot_unmodified_peptides}{If TRUE, it will show the Unmodified
peptides.}

\item{log_base}{The logarithmic scale for the intensity. Default is 2.}

\item{aggregate_PTMs}{If TRUE, same PTM that occur multiple times in the
same peptides,  will be aggregated together.}

\item{palette}{The palette from the Package RColorBrewer. By default is
'Set2'.}

\item{plots_per_page}{Establish the maximum number of plots per page.}
}
\value{
Two plots per sample
}
\description{
Post Translational Modifications
}
\examples{
MQPathCombined <- system.file("extdata/combined/", package = "MQmetrics")
MQCombined <- make_MQCombined(MQPathCombined)
PlotPTM(MQCombined)
}
