% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTM.R
\name{PlotPTM}
\alias{PlotPTM}
\title{Post Translational Modifications}
\usage{
PlotPTM(modificationSpecificPeptides, peptides_modified = 5, palette = "Set2")
}
\arguments{
\item{modificationSpecificPeptides}{modificationSpecificPeptides table from MaxQuant ouput.}

\item{peptides_modified}{Minimum number of peptides modified. Default  is 5.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Two plots per sample
}
\description{
Post Translational Modifications
}
\examples{
MQPathCombined <- system.file('extdata', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
modificationSpecificPeptides <- files[["modificationSpecificPeptides.txt"]]
PlotPTM(modificationSpecificPeptides)

}
