% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPTMAcrossSamples.R
\name{PlotPTMAcrossSamples}
\alias{PlotPTMAcrossSamples}
\title{Title}
\usage{
PlotPTMAcrossSamples(
  MQCombined,
  PTM_of_interest = "Oxidation (M)",
  log_base = 2,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{plots_per_page}{}
}
\value{

}
\description{
Title
}
