% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTotalIonCurrent.R
\name{PlotTotalIonCurrent}
\alias{PlotTotalIonCurrent}
\title{Total Ion Current}
\usage{
PlotTotalIonCurrent(
  msmsScans,
  show_max_value = TRUE,
  palette = "Set2",
  plots_per_page = 5
)
}
\arguments{
\item{msmsScans}{The msmsScans.txt table from  MaxQuant Output.}

\item{show_max_value}{If TRUE, it will show the max TIC value of each sample.}

\item{palette}{The palette from the Package RColorBrewer. By default is 'Set2'.}
}
\value{
Returns a plot the Total Ion Current in each sample. The maximum value is also plotted.
}
\description{
Total Ion Current
}
\examples{
MQPathCombined <- system.file('extdata/combined/', package = 'MQmetrics')
files <- ReadDataFromDir(MQPathCombined)
msmsScans <- files[['msmsScans.txt']]
PlotTotalIonCurrent(msmsScans)


}
