% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTotalIonCurrent.R
\name{PlotTotalIonCurrent}
\alias{PlotTotalIonCurrent}
\title{Plot Total Ion Current}
\usage{
PlotTotalIonCurrent(msmsScans)
}
\arguments{
\item{msmsScans}{The msmsScans.txt table from  MaxQuant Output.}
}
\value{
Returns a plot the Total Ion Current in each sample. The maximum value is also plotted.
}
\description{
Plot Total Ion Current
}
