% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTables.R
\name{ReportTables}
\alias{ReportTables}
\title{Title}
\usage{
ReportTables(MQPathCombined, log_base = 2, intensity_type = "Intensity")
}
\arguments{
\item{intensity_type}{}
}
\value{

}
\description{
Title
}
