% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTables.R
\name{ReportTables}
\alias{ReportTables}
\title{Report Tables with summary data}
\usage{
ReportTables(MQPathCombined, log_base = 2, intensity_type = "Intensity")
}
\arguments{
\item{MQPathCombined}{The directory to the "combined" folder where the MaxQuant results are stored.}

\item{log_base}{The logarithmic scale for the intensity. Default is 2.}

\item{intensity_type}{The type of intensity. Values: 'Intensity' or 'LFQ'.}
}
\value{
A list with four tables are generated:
 - Protein Information
 - Intensity Information
 - Peptide Charge Information
 - Peptide hydrophobicity Information
}
\description{
Report Tables with summary data
}
\examples{
MQPathCombined <- '/home/alvaro/Documents/MaxQuant/example4/'
ReportTables(MQPathCombined)

}
